:sparkles: make certified

+++ debian/setup-certified.sh
@@ -5,6 +5,16 @@ WHOAMI=`python -c 'import os, sys; print os.path.realpath(sys.argv[1])' $0`
 DEBIAN=`dirname $WHOAMI`
 ROOT=`dirname $DEBIAN`
 
+if [ ! -f /home/vagrant/.gitconfig ]; then
+	# Putting this at the top so we don't fail midway through certified-ca.sh setup
+	echo "This setup script uses your .gitconfig to autodetect your email address."
+	echo "Please run:"
+	echo
+	echo "    git config --global user.email \"you@example.com\""
+	echo "    git config --global user.name \"Your Name\""
+	exit 1
+fi
+
 sudo apt-get update
 sudo apt-get -y upgrade
 sudo apt-get -y install ruby-dev
